login-offline: add UserUtils class
