+++ core/src/test/java/org/hisp/dhis/android/core/data/api/BasicAuthenticatorShould.java
@@ -50,7 +50,7 @@ import okhttp3.mockwebserver.RecordedRequest;
 
 import static okhttp3.Credentials.basic;
 import static org.assertj.core.api.Java6Assertions.assertThat;
-import static org.hisp.dhis.android.core.data.api.ApiUtils.base64;
+import static org.hisp.dhis.android.core.utils.UserUtils.base64;
 import static org.mockito.Mockito.when;
 
 // ToDo: Solve problem with INFO logs from MockWebServer being interpreted as errors in gradle
