+++ core/src/test/java/org/hisp/dhis/android/core/user/UserAuthenticateCallUnitShould.java
@@ -64,7 +64,7 @@ import retrofit2.Response;
 
 import static okhttp3.Credentials.basic;
 import static org.assertj.core.api.Java6Assertions.assertThat;
-import static org.hisp.dhis.android.core.data.api.ApiUtils.base64;
+import static org.hisp.dhis.android.core.utils.UserUtils.base64;
 import static org.junit.Assert.fail;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyString;
